Dive into conditional comments and trace outgoing relations conditionally
Includes a small refactor so that we normalize to a predicates object sooner.

Before, incomingMedia and conditionalCommentConditions were carried
around separately, which is unnecessary. This will also make it easier
to add support for <noscript> tracking later.

Additionally, this fixes a couple of places in getCssRulesByProperty
where the predicates weren't propagated correctly.