normalizeFontWeight: Strip +bolder... before looking up the value in css-font-weight-names
This will not yield the correct result when expanding transitions
and animations, in the event that we have to animate or transition
to or from eg. normal+bolder.

I think that's an extremely rare case, and the bug was also there
before. Let's see if we need to handle it -- I think it will require
delaying the expansion of font-weight ranges to snapToAvailable...