Ensure _text is always available on JavaScript assets.
As of assetgraph 7.1.0 assignment to the .text property keeps the
original source code available as _text. This is immensely useful
but has also opened up an unfortunate inconsistency because _text
is not populated from the rawSrc when such assets are instantiated.

In order to avoid adding special cases to callers try to make _text
uniformly available assetgraph is able to decode the raw buffer.